# go-demo-gin/docs/swagger.yaml
definitions:
  authen.LoginForm:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  controllers.TokenResponse:
    properties:
      token:
        type: string
    type: object
  error.Error:
    properties:
      error:
        additionalProperties:
          type: string
        type: object
    type: object
  error.HTTPError:
    properties:
      message:
        $ref: '#/definitions/error.Error'
      statusCode:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/v1/authen/login:
    post:
      consumes:
      - application/json
      description: Login to system
      parameters:
      - description: Login form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authen.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login
      tags:
      - "\U0001F510Authentication"
...
swagger: "2.0"